Remove unused imports and variables

+++ caratApp/src/main/java/edu/berkeley/cs/amplab/carat/android/storage/SimpleHogBug.java
@@ -225,7 +225,7 @@ public class SimpleHogBug implements Serializable, Comparable<SimpleHogBug> {
 			maxError = ebMax - benefit;
 
 		int min = (int) (benefit / 60);
-		int hours = (int) (min / 60);
+		int hours = min / 60;
 		// correct seconds value will be in benefit
 		benefit -= min * 60;
 		min -= hours * 60;
